Start work on #4
write_makefile() now implements a modified version of the idea
by @krlmlr in #4. At the start of each call to make,
a dummy file is created for every clean target, and
no dirty target has a dummy file. The idea is to allow make
to skip clean targets without going through remake.

Problem: clean targets aren't skipped because they all depend
on the phony target .init. Removing .init allows make to skip
clean targets, but this is not a good solution since we
actually need .init.